Behavioural fixes for Eugene and Polygene
Sync is now applied on a rising edge.
Previously sync was not useful for Eugene as the rising sync edge
needed to happen at exactly the same time as a rising clock edge
for it to have any effect.

The current beat is now output on a rising clock edge, then the index is
advanced.
Previously the beat was advanced then the new beat played which had the
effect of shifting all rhythms back one beat.